Move nn_model directory to algorithm directory
